Add function 'roundTimeDown' and coresponding tests.
roundTimeDown will assist in reconsiling if our microservice has become
out of sync with our cron job basis of running every 5 mins. This function
simply takes a python datetime.now() timestamp and rounds only the minute portion
down to the nearest 5 minute increment and returns the five values of the current
time as a list. The return value has this signature (all as integers)
[year, month, day, hour, minute].