Add a new return statement to generateNewCustomTime function.
The function now has an if/else statement that compares the old and new time
format to ensure that the cron job is not running too fast. If the two times
match then we are running too fast and will return the bool False to indicate
that we should not insert any new data and close the db connection. If they are
not equal then we will return the new time.