Removed function in favor of a similar function that rounds down.
This increment function was in order to take the current time and 'round up'
the minute attribute and any upstream unit in the case the time would rollover.
But this was dropped in favor of a function that would 'round down' the minute
attribute, which achieves the same functionality with alot less complexity.